Fix CI/CD compilation errors across workspace
- Fixed BiTemporal::new() calls to include ReceptionTime argument
- Fixed type mismatches in safety.rs (u64/i64) and tracking.rs (Matrix/[f64;3])
- Added type annotations for Vec in compiler.rs
- Implemented TopologyForecaster trait for TopologyForecast
- Added From<DateTime<Utc>> trait for PrecisionTimestamp
- Fixed Modulation::Manchester struct variant with rate field
- Fixed return type mismatches in vendors.rs (schedule_acquisition, cancel_pat, reset, safe_mode)
- Updated sgp4::Elements::from_tle signature to use byte slices
- Fixed u64 * float multiplication by using integer literal
- Fixed PreemptionReason type confusion by re-exporting from crate root
- Fixed ambiguous numeric type and f64 Ord issues in control_node_placement.rs
- Added cost_model field to RouteState
- Fixed move/borrow errors with String values by cloning
- Fixed thiserror String issues in error enums

+++ acs/src/safety.rs
@@ -122,7 +122,7 @@ impl SafetyInterlock {
 
     /// Check watchdog timeout
     pub fn check_watchdog(&self) -> bool {
-        let last = self.last_heartbeat.load(Ordering::Relaxed);
+        let last = self.last_heartbeat.load(Ordering::Relaxed) as i64;
         let now = Utc::now().timestamp_millis();
         (now - last) < self.watchdog_timeout_ms as i64
     }
